Fix semester id being sent as integer + IAP update

+++ app/src/main/java/com/forcetower/uefs/core/model/service/discipline/ServerStatsDiscipline.kt
@@ -72,7 +72,7 @@ data class DisciplineData(
     @SerializedName("group")
     val disciplineGroup: String,
     @SerializedName("semester")
-    val sagresSemesterId: Int,
+    val sagresSemesterId: Long,
     @SerializedName("semester_name")
     val semesterName: String,
     @SerializedName("teacher")
